feat: rename Feed to KnowledgeToFeed
