--- frontend/app/chat/[chatId]/components/ActionsBar/components/Feed/index.ts
@@ -1,1 +1,0 @@
-export * from "./Feed";
