+++ frontend/app/chat/[chatId]/components/ActionsBar/components/KnowledgeToFeed/index.ts
@@ -1,2 +1,1 @@
-export * from "./ChatInput";
 export * from "./KnowledgeToFeed";
